feat: improve code splitting by introducing subpackage entry points
