+++ .eslintrc
@@ -16,7 +16,7 @@
     "import/no-extraneous-dependencies": [
       "error",
       {
-        "devDependencies": ["test/**", "**/*.test.js", "**/*.test.ts"],
+        "devDependencies": ["test/**", "**/*.test.js", "**/*.test.ts", "tsup.config.ts"],
         "packageDir": ["./"]
       }
     ],
