+++ src/allDialects.ts
@@ -1,18 +1,18 @@
-export { bigquery } from './languages/bigquery/bigquery.formatter.js';
-export { db2 } from './languages/db2/db2.formatter.js';
-export { db2i } from './languages/db2i/db2i.formatter.js';
-export { hive } from './languages/hive/hive.formatter.js';
-export { mariadb } from './languages/mariadb/mariadb.formatter.js';
-export { mysql } from './languages/mysql/mysql.formatter.js';
-export { tidb } from './languages/tidb/tidb.formatter.js';
-export { n1ql } from './languages/n1ql/n1ql.formatter.js';
-export { plsql } from './languages/plsql/plsql.formatter.js';
-export { postgresql } from './languages/postgresql/postgresql.formatter.js';
-export { redshift } from './languages/redshift/redshift.formatter.js';
-export { spark } from './languages/spark/spark.formatter.js';
-export { sqlite } from './languages/sqlite/sqlite.formatter.js';
-export { sql } from './languages/sql/sql.formatter.js';
-export { trino } from './languages/trino/trino.formatter.js';
-export { transactsql } from './languages/transactsql/transactsql.formatter.js';
-export { singlestoredb } from './languages/singlestoredb/singlestoredb.formatter.js';
-export { snowflake } from './languages/snowflake/snowflake.formatter.js';
+export { bigquery } from './languages/bigquery/index.js';
+export { db2 } from './languages/db2/index.js';
+export { db2i } from './languages/db2i/index.js';
+export { hive } from './languages/hive/index.js';
+export { mariadb } from './languages/mariadb/index.js';
+export { mysql } from './languages/mysql/index.js';
+export { tidb } from './languages/tidb/index.js';
+export { n1ql } from './languages/n1ql/index.js';
+export { plsql } from './languages/plsql/index.js';
+export { postgresql } from './languages/postgresql/index.js';
+export { redshift } from './languages/redshift/index.js';
+export { spark } from './languages/spark/index.js';
+export { sqlite } from './languages/sqlite/index.js';
+export { sql } from './languages/sql/index.js';
+export { trino } from './languages/trino/index.js';
+export { transactsql } from './languages/transactsql/index.js';
+export { singlestoredb } from './languages/singlestoredb/index.js';
+export { snowflake } from './languages/snowflake/index.js';
