+++ src/dialect.ts
@@ -5,6 +5,10 @@ import {
 import Tokenizer from './lexer/Tokenizer.js';
 import { TokenizerOptions } from './lexer/TokenizerOptions.js';
 
+import { defaultFormatOptions, FormatOptions } from './FormatOptions.js';
+import Formatter from './formatter/Formatter.js';
+import { validateConfig } from './validateConfig.js';
+
 export interface DialectOptions {
   name: string;
   tokenizerOptions: TokenizerOptions;
@@ -46,3 +50,31 @@ const processDialectFormatOptions = (
     (options.tabularOnelineClauses ?? options.onelineClauses).map(name => [name, true])
   ),
 });
+
+export type FormatOptionsWithDialect = Partial<FormatOptions> & {
+  dialect: DialectOptions;
+};
+
+/**
+ * Formats and SQL query string similar to `format` function, however it works directly with dialect objects instead of dialect names.
+ * Provides better bundling performance and control over the formatting process.
+ *
+ * @param {string} query - The SQL query string to format
+ * @param {FormatOptionsWithDialect} options - Configuration options
+ * @return {string} formatted query
+ */
+export const formatDialect = (
+  query: string,
+  { dialect, ...cfg }: FormatOptionsWithDialect
+): string => {
+  if (typeof query !== 'string') {
+    throw new Error('Invalid query argument. Expected string, instead got ' + typeof query);
+  }
+
+  const options = validateConfig({
+    ...defaultFormatOptions,
+    ...cfg,
+  });
+
+  return new Formatter(createDialect(dialect), options).format(query);
+};
