+++ src/languages/bigquery/index.ts
@@ -1,0 +1,1 @@
+export { bigquery } from './bigquery.formatter.js';
