+++ src/languages/db2/index.ts
@@ -1,0 +1,1 @@
+export { db2 } from './db2.formatter.js';
