+++ src/languages/db2i/index.ts
@@ -1,0 +1,1 @@
+export { db2i } from './db2i.formatter.js';
