+++ src/languages/hive/index.ts
@@ -1,0 +1,1 @@
+export { hive } from './hive.formatter.js';
