+++ src/languages/mariadb/index.ts
@@ -1,0 +1,1 @@
+export { mariadb } from './mariadb.formatter.js';
