+++ src/languages/mysql/index.ts
@@ -1,0 +1,1 @@
+export { mysql } from './mysql.formatter.js';
