+++ src/languages/n1ql/index.ts
@@ -1,0 +1,1 @@
+export { n1ql } from './n1ql.formatter.js';
