+++ src/languages/plsql/index.ts
@@ -1,0 +1,1 @@
+export { plsql } from './plsql.formatter.js';
