+++ src/languages/postgresql/index.ts
@@ -1,0 +1,1 @@
+export { postgresql } from './postgresql.formatter.js';
