+++ src/languages/redshift/index.ts
@@ -1,0 +1,1 @@
+export { redshift } from './redshift.formatter.js';
