+++ src/languages/singlestoredb/index.ts
@@ -1,0 +1,1 @@
+export { singlestoredb } from './singlestoredb.formatter.js';
