+++ src/languages/snowflake/index.ts
@@ -1,0 +1,1 @@
+export { snowflake } from './snowflake.formatter.js';
