+++ src/languages/spark/index.ts
@@ -1,0 +1,1 @@
+export { spark } from './spark.formatter.js';
