+++ src/languages/sql/index.ts
@@ -1,0 +1,1 @@
+export { sql } from './sql.formatter.js';
