+++ src/languages/sqlite/index.ts
@@ -1,0 +1,1 @@
+export { sqlite } from './sqlite.formatter.js';
