+++ src/languages/tidb/index.ts
@@ -1,0 +1,1 @@
+export { tidb } from './tidb.formatter.js';
