+++ src/languages/transactsql/index.ts
@@ -1,0 +1,1 @@
+export { transactsql } from './transactsql.formatter.js';
