+++ src/languages/trino/index.ts
@@ -1,0 +1,1 @@
+export { trino } from './trino.formatter.js';
