+++ test/sqlFormatter.test.ts
@@ -1,6 +1,8 @@
 import dedent from 'dedent-js';
 
-import { format, formatDialect, SqlLanguage, sqlite, DialectOptions } from '../src/index.js';
+import { format, SqlLanguage } from '../src/index.js';
+import { formatDialect, DialectOptions } from '../src/dialect.js';
+import { sqlite } from '../src/languages/sqlite/index.js';
 
 describe('sqlFormatter', () => {
   it('throws error when unsupported language parameter specified', () => {
